+++ source/Model/Bezier.ts
@@ -1,0 +1,8 @@
+import { Point } from "./Point";
+
+export interface Bezier {
+    from: Point,
+    to: Point,
+    control1: Point,
+    control2: Point
+}